+++ app/assets/data/xk/localization.ts
@@ -1,0 +1,25 @@
+const localizationkosovo = [
+  {
+    name: 'Kosovo',
+    countryId: 21,
+    displayName: 'Kosovo (Kosova)',
+    languages: [
+      {
+        name: 'Albanian',
+        displayName: 'Shqip',
+        languageCode: 'xk-sq',
+        locale: 'XKsq',
+        luxonLocale:'sq-XK'
+      },
+      {
+        name: 'Serbian',
+        displayName: 'Srpski',
+        languageCode: 'xk-rs',
+        locale: 'XKrs',
+        luxonLocale:'sr-Latn-XK'
+      },
+    ],
+  }
+];
+export default localizationkosovo;
+
