+++ app/assets/translations/appConstants/bebbo/MKsq.ts
