+++ app/assets/translations/appConstants/xk/XKsq.ts
