+++ app/assets/translations/appOfflineData/apiConstants.ts
@@ -3,8 +3,8 @@ import RNFS from 'react-native-fs';
 export const destinationFolder=RNFS.DocumentDirectoryPath + '/content/';
 export const buildForFoleja = 'foleja'; //'foleja'
 export const buildForBebbo = 'bebbo'; //'foleja'
-export const buildFor = buildForFoleja; //'foleja'
-//export const buildFor = buildForBebbo; //'foleja'
+//export const buildFor = buildForFoleja; //'foleja'
+export const buildFor = buildForBebbo; //'foleja'
 export const maxRelatedArticleSize = 3;
 export const isArticlePinned = '1';
 export const articleCategory = '4,1,55,56,3,2';
