+++ app/dynamicImports.ts
@@ -1,0 +1,20 @@
+require('intl/locale-data/jsonp/sq-XK')
+require('intl/locale-data/jsonp/sr-Latn-XK')
+export const localization=require('@assets/data/xk/localization').default;
+export const bebbo_logo_shape=require('@assets/svg/xk/xk_logo_shape.svg');
+export const img_logo_chatbot_new=require('@assets/svg/xk/img-logo-chatbot-new.svg');
+export const AVAILABLE_LANGUAGES =require('../../../localization/Available_lang_xk').default;
+export const taxonomydata =require('@offlinedata/xk/taxonomies').taxonomydata;
+export const ActivitiesData =require('@offlinedata/xk/ActivitiesData').ActivitiesData;
+export const articledata = require('@offlinedata/xk/article').articledata;
+export const basicPagesData =require('@offlinedata/xk/basicPages').basicPagesData;
+export const ChildDevelopmentData= require('@offlinedata/xk/ChildDevelopmentData').ChildDevelopmentData;
+export const FaqsData = require('@offlinedata/xk/FaqsData').FaqsData ;
+export const healthCheckupsData =require('@offlinedata/xk/healthCheckupsData').healthCheckupsData;
+export const dailyHomeNotificationdata =require('@offlinedata/xk/dailyHomeNotification').dailyHomeNotificationdata;
+export const MileStonesData =require('@offlinedata/xk/MileStonesData').MileStonesData;
+export const PinnedChildDevData =require('@offlinedata/xk/PinnedChildDevData').PinnedChildDevData;
+export const standardDevData =require('@offlinedata/xk/standardDevData').standardDevData;
+export const SurveyData =require('@offlinedata/xk/SurveyData').SurveyData;
+export const vaccineData =require('@offlinedata/xk/vaccineData').vaccineData;
+export const VideoArticleData =require('@offlinedata/xk/VideoArticleData').VideoArticleData;
