+++ app/dynamicImportsbebbo.ts
@@ -1,0 +1,35 @@
+require('intl/locale-data/jsonp/sq-AL')
+require('intl/locale-data/jsonp/be-BY')
+require('intl/locale-data/jsonp/ru-BY')
+require('intl/locale-data/jsonp/el-GR')
+require('intl/locale-data/jsonp/ky-KG')
+require('intl/locale-data/jsonp/ru-KG')
+require('intl/locale-data/jsonp/mk-MK')
+require('intl/locale-data/jsonp/sq-MK')
+require('intl/locale-data/jsonp/ar')
+require('intl/locale-data/jsonp/fa-AF')
+require('intl/locale-data/jsonp/sr-Latn-RS')
+require('intl/locale-data/jsonp/uz-Latn-UZ')
+require('intl/locale-data/jsonp/ru-RU')
+require('intl/locale-data/jsonp/en-US')
+require('@assets/customLocale/bg-BG')
+require('@assets/customLocale/tg-TJ')
+require('@assets/customLocale/me-cnr')
+export const localization=require('@assets/data/bebbo/localization').default;
+export const bebbo_logo_shape=require('@assets/svg/bebbo/bebbo_logo_shape.svg');
+export const img_logo_chatbot_new=require('@assets/svg/bebbo/img-logo-chatbot-new.svg');
+export const AVAILABLE_LANGUAGES =require('../../../localization/Available_lang_bebbo').default;
+export const taxonomydata =require('@offlinedata/bebbo/taxonomies').taxonomydata;
+export const ActivitiesData =require('@offlinedata/bebbo/ActivitiesData').ActivitiesData;
+export const articledata = require('@offlinedata/bebbo/article').articledata;
+export const basicPagesData =require('@offlinedata/bebbo/basicPages').basicPagesData;
+export const ChildDevelopmentData= require('@offlinedata/bebbo/ChildDevelopmentData').ChildDevelopmentData;
+export const FaqsData = require('@offlinedata/bebbo/FaqsData').FaqsData ;
+export const healthCheckupsData =require('@offlinedata/bebbo/healthCheckupsData').healthCheckupsData;
+export const dailyHomeNotificationdata =require('@offlinedata/bebbo/dailyHomeNotification').dailyHomeNotificationdata;
+export const MileStonesData =require('@offlinedata/bebbo/MileStonesData').MileStonesData;
+export const PinnedChildDevData =require('@offlinedata/bebbo/PinnedChildDevData').PinnedChildDevData;
+export const standardDevData =require('@offlinedata/bebbo/standardDevData').standardDevData;
+export const SurveyData =require('@offlinedata/bebbo/SurveyData').SurveyData;
+export const vaccineData =require('@offlinedata/bebbo/vaccineData').vaccineData;
+export const VideoArticleData =require('@offlinedata/bebbo/VideoArticleData').VideoArticleData;
