+++ app/localization/Available_lang_bebbo.ts
@@ -1,21 +1,21 @@
-import ALsq from "@assets/translations/appConstants/ALsq";
-import BGbg from "@assets/translations/appConstants/BGbg";
-import BYbe from "@assets/translations/appConstants/BYbe";
-import BYru from "@assets/translations/appConstants/BYru";
-import en from "@assets/translations/appConstants/en";
-import GRel from "@assets/translations/appConstants/GRel";
-import KGky from "@assets/translations/appConstants/KGky";
-import KGru from "@assets/translations/appConstants/KGru";
-import MEcnr from "@assets/translations/appConstants/MEcnr";
-import MKmk from "@assets/translations/appConstants/MKmk";
-import MKsq from "@assets/translations/appConstants/MKsq";
-import RSen from "@assets/translations/appConstants/RSen";
-import RSsr from "@assets/translations/appConstants/RSsr";
-import ru from "@assets/translations/appConstants/ru";
-import TJru from "@assets/translations/appConstants/TJru";
-import TJtg from "@assets/translations/appConstants/TJtg";
-import UZru from "@assets/translations/appConstants/UZru";
-import UZuz from "@assets/translations/appConstants/UZuz";
+import ALsq from "@assets/translations/appConstants/bebbo/ALsq";
+import BGbg from "@assets/translations/appConstants/bebbo/BGbg";
+import BYbe from "@assets/translations/appConstants/bebbo/BYbe";
+import BYru from "@assets/translations/appConstants/bebbo/BYru";
+import en from "@assets/translations/appConstants/bebbo/en";
+import GRel from "@assets/translations/appConstants/bebbo/GRel";
+import KGky from "@assets/translations/appConstants/bebbo/KGky";
+import KGru from "@assets/translations/appConstants/bebbo/KGru";
+import MEcnr from "@assets/translations/appConstants/bebbo/MEcnr";
+import MKmk from "@assets/translations/appConstants/bebbo/MKmk";
+import MKsq from "@assets/translations/appConstants/bebbo/MKsq";
+import RSen from "@assets/translations/appConstants/bebbo/RSen";
+import RSsr from "@assets/translations/appConstants/bebbo/RSsr";
+import ru from "@assets/translations/appConstants/bebbo/ru";
+import TJru from "@assets/translations/appConstants/bebbo/TJru";
+import TJtg from "@assets/translations/appConstants/bebbo/TJtg";
+import UZru from "@assets/translations/appConstants/bebbo/UZru";
+import UZuz from "@assets/translations/appConstants/bebbo/UZuz";
 
 const AVAILABLE_LANGUAGES = {
     en,
