+++ app/navigation/LocalizationNavigation.tsx
@@ -1,4 +1,4 @@
-import localization from '@assets/data/localization';
+import  {localization}  from '../dynamicImports';
 import { createStackNavigator } from '@react-navigation/stack';
 import CountryLanguageConfirmation from '@screens/localization/CountryLanguageConfirmation';
 import CountrySelection from '@screens/localization/CountrySelection';
