+++ app/redux/reducers/localizationSlice.ts
@@ -1,4 +1,4 @@
-import localization from '@assets/data/localization';
+import {localization} from '../../dynamicImports';
 import { createSlice, PayloadAction } from '@reduxjs/toolkit';
 // import {RootState} from './../../../App';
