+++ babel.config.js
@@ -43,7 +43,6 @@ module.exports = (api) => {
   if (babelEnv !== 'development') {
     plugins.push(['transform-remove-console', {exclude: ['error', 'warn']}]);
   }
-  // plugins.push(["react-native-reanimated/plugin"]);
   return {
     presets: ['module:metro-react-native-babel-preset'],
     plugins: plugins,
